Refactor: [모집완료] 상태 변경 메소드 meeting.service 도메인 패키지 이동
