+++ src/main/java/com/boardgo/domain/meeting/service/MeetingCommandServiceV1.java
@@ -1,6 +1,8 @@
 package com.boardgo.domain.meeting.service;
 
-import static com.boardgo.common.constant.S3BucketConstant.*;
+import static com.boardgo.common.constant.S3BucketConstant.MEETING;
+import static com.boardgo.domain.meeting.entity.enums.MeetingState.COMPLETE;
+import static com.boardgo.domain.meeting.entity.enums.MeetingState.PROGRESS;
 
 import com.boardgo.common.exception.CustomNoSuchElementException;
 import com.boardgo.common.utils.FileUtils;
@@ -12,6 +14,7 @@ import com.boardgo.domain.mapper.MeetingMapper;
 import com.boardgo.domain.meeting.controller.request.MeetingCreateRequest;
 import com.boardgo.domain.meeting.entity.MeetingEntity;
 import com.boardgo.domain.meeting.repository.MeetingRepository;
+import java.util.List;
 import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -66,5 +69,12 @@ public class MeetingCommandServiceV1 implements MeetingCommandUseCase {
             imageUri = s3Service.upload(MEETING, FileUtils.getUniqueFileName(imageFile), imageFile);
         }
         return imageUri;
+    }
+
+    @Override
+    public void updateCompleteMeetingState() {
+        List<Long> meetingIds = meetingRepository.findCompleteMeetingId(PROGRESS);
+        List<MeetingEntity> meetingEntities = meetingRepository.findByIdIn(meetingIds);
+        meetingEntities.forEach((meetingEntity -> meetingEntity.updateMeetingState(COMPLETE)));
     }
 }
