+++ src/test/java/com/boardgo/integration/meeting/service/MeetingBatchServiceV1Test.java
@@ -1,11 +1,7 @@
 package com.boardgo.integration.meeting.service;
 
-import static com.boardgo.domain.meeting.entity.enums.MeetingState.COMPLETE;
 import static com.boardgo.domain.review.entity.enums.ReviewType.FINISH;
 import static com.boardgo.integration.data.MeetingData.getMeetingEntityData;
-import static com.boardgo.integration.data.UserInfoData.userInfoEntityData;
-import static com.boardgo.integration.fixture.MeetingParticipantFixture.getLeaderMeetingParticipantEntity;
-import static com.boardgo.integration.fixture.MeetingParticipantFixture.getParticipantMeetingParticipantEntity;
 import static com.boardgo.integration.fixture.UserInfoFixture.socialUserInfoEntity;
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -54,43 +50,5 @@ public class MeetingBatchServiceV1Test extends IntegrationTestSupport {
                 meetingEntity -> {
                     assertThat(meetingEntity.getState().toString()).isEqualTo(FINISH.toString());
                 });
-    }
-
-    @Test
-    @DisplayName("모임의 인원이 정원인 경우 모임 완료 처리 된다")
-    void 모임의_인원이_정원인_경우_모임_완료_처리_된다() {
-        // given
-        List<Long> meetingIds = new ArrayList<>();
-        int limit = 5;
-        for (int i = 0; i < 10; i++) {
-            userRepository.save(
-                    userInfoEntityData()
-                            .email(i + "email@naver.com")
-                            .nickName("내이름" + i)
-                            .providerType(ProviderType.LOCAL)
-                            .build());
-            MeetingEntity meeting = getMeetingEntityData(1L).limitParticipant(limit).build();
-            meetingIds.add(meetingRepository.save(meeting).getId());
-            meetingParticipantRepository.save(
-                    getLeaderMeetingParticipantEntity(meeting.getId(), 1L));
-        }
-        // 모임정원
-        for (int i = 1; i < limit; i++) {
-            for (int j = i; j < limit; j++) {
-                meetingParticipantRepository.save(
-                        getParticipantMeetingParticipantEntity((long) i, (long) j));
-            }
-        }
-
-        // when
-        meetingBatchService.updateCompleteMeetingState();
-
-        // then
-        List<MeetingEntity> meetingEntities = meetingRepository.findByIdIn(meetingIds);
-        int index = 0;
-        while (limit == index) {
-            MeetingEntity meetingEntity = meetingEntities.get(index);
-            assertThat(meetingEntity.getState().toString()).isEqualTo(COMPLETE.toString());
-        }
     }
 }
