+++ src/test/java/com/boardgo/integration/meeting/service/MeetingCommandServiceV1Test.java
@@ -1,12 +1,24 @@
 package com.boardgo.integration.meeting.service;
 
+import static com.boardgo.domain.meeting.entity.enums.MeetingState.COMPLETE;
+import static com.boardgo.integration.data.MeetingData.getMeetingEntityData;
+import static com.boardgo.integration.data.UserInfoData.userInfoEntityData;
+import static com.boardgo.integration.fixture.MeetingParticipantFixture.getLeaderMeetingParticipantEntity;
+import static com.boardgo.integration.fixture.MeetingParticipantFixture.getParticipantMeetingParticipantEntity;
+import static org.assertj.core.api.Assertions.assertThat;
+
 import com.boardgo.domain.meeting.entity.MeetingEntity;
 import com.boardgo.domain.meeting.entity.enums.MeetingType;
+import com.boardgo.domain.meeting.repository.MeetingParticipantRepository;
 import com.boardgo.domain.meeting.repository.MeetingRepository;
 import com.boardgo.domain.meeting.service.MeetingCommandUseCase;
+import com.boardgo.domain.user.entity.enums.ProviderType;
+import com.boardgo.domain.user.repository.UserRepository;
 import com.boardgo.integration.fixture.MeetingFixture;
 import com.boardgo.integration.support.IntegrationTestSupport;
 import jakarta.persistence.EntityManager;
+import java.util.ArrayList;
+import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -14,6 +26,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 public class MeetingCommandServiceV1Test extends IntegrationTestSupport {
     @Autowired private MeetingRepository meetingRepository;
+    @Autowired private UserRepository userRepository;
+    @Autowired private MeetingParticipantRepository meetingParticipantRepository;
     @Autowired private MeetingCommandUseCase meetingCommandUseCase;
     @Autowired private EntityManager entityManager;
 
@@ -31,5 +45,43 @@ public class MeetingCommandServiceV1Test extends IntegrationTestSupport {
         MeetingEntity meeting = meetingRepository.findById(savedMeeting.getId()).get();
         System.out.println("meeting = " + meeting);
         Assertions.assertThat(meeting.getShareCount()).isEqualTo(1L);
+    }
+
+    @Test
+    @DisplayName("모임의 인원이 정원인 경우 모임 완료 처리 된다")
+    void 모임의_인원이_정원인_경우_모임_완료_처리_된다() {
+        // given
+        List<Long> meetingIds = new ArrayList<>();
+        int limit = 5;
+        for (int i = 0; i < 10; i++) {
+            userRepository.save(
+                    userInfoEntityData()
+                            .email(i + "email@naver.com")
+                            .nickName("내이름" + i)
+                            .providerType(ProviderType.LOCAL)
+                            .build());
+            MeetingEntity meeting = getMeetingEntityData(1L).limitParticipant(limit).build();
+            meetingIds.add(meetingRepository.save(meeting).getId());
+            meetingParticipantRepository.save(
+                    getLeaderMeetingParticipantEntity(meeting.getId(), 1L));
+        }
+        // 모임정원
+        for (int i = 1; i < limit; i++) {
+            for (int j = i; j < limit; j++) {
+                meetingParticipantRepository.save(
+                        getParticipantMeetingParticipantEntity((long) i, (long) j));
+            }
+        }
+
+        // when
+        meetingCommandUseCase.updateCompleteMeetingState();
+
+        // then
+        List<MeetingEntity> meetingEntities = meetingRepository.findByIdIn(meetingIds);
+        int index = 0;
+        while (limit == index) {
+            MeetingEntity meetingEntity = meetingEntities.get(index);
+            assertThat(meetingEntity.getState().toString()).isEqualTo(COMPLETE.toString());
+        }
     }
 }
